App ver 3.0

+++ ModuleHandlerDataFromUser/src/main/java/app/data/DataFromWebsocket.java
@@ -11,6 +11,6 @@ public class DataFromWebsocket {
     private String param;
     private int stage;
     private boolean control;
-    private int minValue;
-    private int maxValue;
+    private float minValue;
+    private float maxValue;
 }
